requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=1.0.0
boto3>=1.34.0
opensearch-py>=2.4.0
requests-aws4auth>=1.3.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
langchain-aws>=0.1.0
